Organize commands into modular files
